+++ packages/components/vite.config.ts
@@ -35,6 +35,7 @@ export default defineConfig(configEnv => {
             'lodash': 'lodash',
             'react-is': 'react-is',
             '@iconify/react': '@iconify/react',
+            'react-router-dom': "react-router-dom"
           }
         }
       },